+++ packages/stubee2-rolling-design-core/src/RollingPalette.ts
@@ -1,42 +1,43 @@
 const text = {
-  Ligther: "rgba(114, 119, 122, 1)",
-  Light: "rgba(108, 112, 114, 1)",
-  Base: "rgba(64, 68, 70, 1)",
-  Dark: "rgba(48, 52, 55, 1)",
-  Darker: "rgba(32, 35, 37, 1)",
-  Darkest: "rgba(9, 10, 10, 1)",
+  Ligther: "#72777A",
+  Light: "#6C7072",
+  Base: "#404446",
+  Dark: "#303437",
+  Darker: "#202325",
+  Darkest: "#090A0A",
 };
 
 const unEmphasize = {
-  Lightest: "rgba(247, 249, 250, 1)",
-  Lighter: "rgba(242, 244, 245, 1)",
-  Light: "rgba(115, 123, 152, 0)",
-  Base: "rgba(189, 194, 208, 1)",
-  Dark: "rgba(115, 123, 152, 1)",
+  Lightest: "#F7F9FA",
+  Lighter: "#F2F4F5",
+  Light: "#DEE0E6",
+  Base: "#BDC2D0",
+  Dark: "#737B98",
 };
 
 const main = {
-  Lightest: "rgba(255, 199, 196, 1)",
-  Lighter: "rgba(153, 179, 239, 1)",
-  Light: "rgba(72, 105, 246, 1)",
-  Dark: "rgba(29, 30, 90, 1)",
-  Darkest: "rgba(8, 12, 43, 1)",
+  Lightest: "#D5E4F2",
+  Lighter: "#99B3EF",
+  Light: "#5667C1",
+  Base: "#4869F6",
+  Dark: "#1C1F5B",
+  Darkest: "#080C2B",
 };
 
 const error = {
-  Lightest: "rgba(255, 199, 196, 1)",
-  Lighter: "rgba(255, 145, 139, 1)",
-  Light: "rgba(236, 106, 94, 1)",
-  Base: "rgba(241, 40, 28, 1)",
-  Dark: "rgba(214, 21, 9, 1)",
+  Lightest: "#FFC7C4",
+  Lighter: "#FF918B",
+  Light: "#FF4E43",
+  Base: "#F1281C",
+  Dark: "#D61509",
 };
 
 const check = {
-  Lightest: "rgba(236, 252, 229, 1)",
-  Lighter: "rgba(125, 222, 134, 1)",
-  Light: "rgba(76, 212, 113, 1)",
-  Base: "rgba(23, 198, 102, 1)",
-  Dark: "rgba(25, 129, 85, 1)",
+  Lightest: "#ECFCE5",
+  Lighter: "#7DDE86",
+  Light: "#4CD471",
+  Base: "#17C666",
+  Dark: "#198155",
 };
 
 export const RollingPalette = {