+++ packages/stubee2-rolling-ui/src/components/Layout/DropDownSelect/type.ts
@@ -4,8 +4,6 @@ export interface DrowDownSelectProps {
   dropDownSelectItems: string[];
   width: string;
   height: string;
-  isError?: boolean;
-  errorMessage?: string;
   selectItem: string;
   setSelectItem: Dispatch<SetStateAction<string>>;
 }