+++ packages/stubee2-rolling-ui/src/components/Layout/TextInput/style.ts
@@ -4,21 +4,25 @@ export const TextInputBar = styled.input<{
   customStyle?: CSSObject;
   isError?: boolean;
 }>`
-  ${({ customStyle }) => customStyle}
   border: ${({ isError }) =>
     isError
-      ? "1px solid rgba(241, 40, 28, 1)"
-      : "1px solid rgba(189, 194, 208, 1);"};
+      ? "1px solid rgba(255, 78, 67, 1)"
+      : " 1px solid rgba(189, 194, 208, 1);"};
   border-radius: 6px;
   outline: none;
   resize: none;
   ::placeholder {
     color: rgba(189, 194, 208, 1);
   }
+  ${({ customStyle }) => customStyle}
 `;
 
 export const TextInputContainer = styled.div`
   display: flex;
   flex-direction: column;
   row-gap: 5px;
+`;
+
+export const ErrorText = styled.p`
+  color: rgba(255, 78, 67, 1);
 `;