+++ packages/stubee2-rolling-ui/src/index.ts
@@ -1,4 +1,2 @@
-export * from "./components/Button/Button";
-export * from "./components/TextInput/TextInput";
-export * from "./components/StarRating/StarRating";
-export * from "./components/DropDownSelect/DropDownSelect";
+export * from "./components/Layout";
+export * from "./components/Ui";