+++ packages/stubee2-rolling-design-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @stubee2/stubee2-rolling-design-token
 
+## 1.2.1
+
+### Patch Changes
+
+- fix 401 error
+
+## 1.2.0
+
+### Minor Changes
+
+- change rgb to hex
+
 ## 0.1.0
 
 ### Minor Changes